insights-api/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==39.0.2
curlify==2.2.1
facebook-business==16.0.0
Flask==2.2.3
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
importlib-metadata==6.0.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
mysql-connector-python==8.0.32
protobuf==3.20.3
pycountry==22.3.5
pycparser==2.21
PyJWT==1.7.1
PyMySQL==1.0.2
python-dotenv==0.21.1
PyYAML==6.0
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.4
typing-extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
yarl==1.8.2
zipp==3.14.0
Flask-Cors==3.0.10
